LinkTo command works :grin:

+++ Toffee/NetFxCsproj.cs
@@ -2,12 +2,16 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using Toffee.Infrastructure;
 
 namespace Toffee
 {
     class NetFxCsproj : INetFxCsproj
     {
+        private const string Indentation6 = "      ";
+        private const string Indentation4 = "    ";
+
         private readonly IFilesystem _filesystem;
 
         public NetFxCsproj(IFilesystem filesystem)
@@ -15,45 +19,95 @@ namespace Toffee
             _filesystem = filesystem;
         }
 
-        public Dictionary<string, string> ReplaceReferencedNuGetDllsWithLinkDlls(string csprojPath, Link link, string[] dlls)
+        public IReadOnlyCollection<ReplacementRecord> ReplaceReferencedNuGetDllsWithLinkDlls(string csprojPath, Link link, string[] dlls)
         {
-            var replacedLines = new Dictionary<string, string>();
+            var replacedLines = new List<ReplacementRecord>();
 
             var lines = _filesystem.ReadAllLines(csprojPath).ToArray();
 
             var linesCopy = new List<string>();
-
+            
             for (var i = 0; i < lines.Length; i++)
             {
                 var line = lines[i];
-                
-                (var isReference, var dll) = IsReferenceToInstalledNuGetPackage(line, dlls);
 
-                if (isReference)
+                if (IsEndOfFile(i, lines))
                 {
                     linesCopy.Add(line);
+                    continue;
+                }
 
-                    var commentLine = $"<!-- Line below was replaced by Toffee at {DateTime.Now:dd.MM.yyyy HH:mm:ss}. Original line: {line} -->";
-                    linesCopy.Add(commentLine);
+                var nextLine = lines[i + 1];
 
-                    var replacedLine = ConstructHintPathLineWithLinkedDll(link, dll);
-                    linesCopy.Add(replacedLine);
+                (var isReference, var dll) = IsReferenceToInstalledNuGetPackage(line, dlls);
 
-                    replacedLines.Add(line, replacedLine);
+                var isNextLineHintPath = IsNextLineHintPathPointingToNuGetDll(nextLine);
 
-                    i = i + 3; // Skipping next line (comment line), and the one after (HintPath line, now replaced by new dll path)
+                if (isReference && isNextLineHintPath)
+                {
+                    var replacementRecord = new ReplacementRecord();
+
+                    // Write comment above the <Reference Inclue=""> line
+                    var commentForReplacedReferenceLine = $"{Indentation4}<!-- Line below was replaced by Toffee at {DateTime.Now:dd.MM.yyyy HH:mm:ss}. Original line: {line.Trim()} -->";
+                    linesCopy.Add(commentForReplacedReferenceLine);
+
+                    // Replace the <Reference Include=""> line with assembly info for the linked dll instead of the original dll
+                    var replacedReferenceLine = ConstructReferenceLineWithLinkedDllAssemblyInfo(link, dll);
+                    linesCopy.Add(replacedReferenceLine);
+
+                    replacementRecord.OriginalReferenceElement = line.Trim();
+                    replacementRecord.NewReferenceElement = replacedReferenceLine.Trim();
+
+                    // Write comment above the next <HintPath> line
+                    var commentForReplacedHintPathLine = $"{Indentation6}<!-- Line below was replaced by Toffee at {DateTime.Now:dd.MM.yyyy HH:mm:ss}. Original line: {nextLine.Trim()} -->";
+                    linesCopy.Add(commentForReplacedHintPathLine);
+
+                    // Replace the next <HintPath>{path}</HintPath> line with path to local dll instead of nuget dll
+                    var replacedHintPathLine = ConstructHintPathLineWithLinkedDll(link, dll);
+                    linesCopy.Add(replacedHintPathLine);
+
+                    replacementRecord.OriginalHintPathElement = nextLine.Trim();
+                    replacementRecord.NewHintPathElement = replacedHintPathLine.Trim();
+                    
+                    replacedLines.Add(replacementRecord);
+
+                    i++;
                 }
                 else
                 {
                     linesCopy.Add(line);
-
-                    i++;
                 }
             }
 
-            _filesystem.WriteAllLines(csprojPath, linesCopy);
+            if (replacedLines.Any())
+            {
+                _filesystem.WriteAllLines(csprojPath, linesCopy);
+            }
 
             return replacedLines;
+        }
+
+        private string ConstructReferenceLineWithLinkedDllAssemblyInfo(Link link, string dll)
+        {
+            var linkDllPath = ConstructPathToLinkDll(link, dll);
+
+            if (_filesystem.FileExists(linkDllPath))
+            {
+                var linkedAssembly = Assembly.LoadFile(linkDllPath);
+                return $"{Indentation4}<Reference Include=\"{linkedAssembly.FullName}\">";
+            }
+
+            throw new FileNotFoundException($"The constructed path to linked dll \"{linkDllPath}\" does not exist");
+        }
+
+        private static bool IsEndOfFile(int i, string[] lines)
+        {
+            return i + 1 >= lines.Length;
+        }
+
+        private static bool IsNextLineHintPathPointingToNuGetDll(string nextLine)
+        {
+            return nextLine.TrimStart().StartsWith("<HintPath>..\\packages\\");
         }
 
         private string ConstructHintPathLineWithLinkedDll(Link link, string dll)
@@ -62,7 +116,7 @@ namespace Toffee
 
             if (_filesystem.FileExists(linkDllPath))
             {
-                return $"<HintPath>{linkDllPath}</HintPath>";
+                return $"{Indentation6}<HintPath>{linkDllPath}</HintPath>";
             }
 
             throw new FileNotFoundException($"The constructed path to linked dll \"{linkDllPath}\" does not exist");
@@ -77,7 +131,9 @@ namespace Toffee
         {
             foreach (var dll in dlls)
             {
-                if (line.StartsWith($"<Reference Include=\"{dll}, Version"))
+                var s = string.Format("<Reference Include=\"{0}, Version=", dll);
+
+                if (line.TrimStart().StartsWith(s))
                 {
                     return (true, dll);
                 }
